+++ CHANGELOG.md
@@ -50,7 +50,7 @@
   * Added new functions related to single document operations:
     * `ElasticsearchEx.Api.Document.index/2`
     * `ElasticsearchEx.Api.Document.create/2`
-    * `ElasticsearchEx.Api.Document.get_document/11 to Api.Document.get/11`
+    * `ElasticsearchEx.Api.Document.get_document/1`
     * `ElasticsearchEx.Api.Document.get_source/1`
     * `ElasticsearchEx.Api.Document.document_exists?/1`
     * `ElasticsearchEx.Api.Document.source_exists?/1`