+++ CHANGELOG.md
@@ -5,9 +5,9 @@
 
 * **Changes:**
   * Renamed `ElasticsearchEx.Api.Document.get_document/1` to `ElasticsearchEx.Api.Document.get/1`
-  * Replaced `ElasticsearchEx.Api.Document.get_source/1` by `ElasticsearchEx.Api.Document.get/1` with `source_only: true` option
+  * Moved `ElasticsearchEx.Api.Document.get_source/1` to `ElasticsearchEx.Api.Source.get/1`
   * Renamed `ElasticsearchEx.Api.Document.document_exists?/1` to `ElasticsearchEx.Api.Document.exists?/1`
-  * Replaced `ElasticsearchEx.Api.Document.source_exists?/1` by `ElasticsearchEx.Api.Document.exists?/1` with `source_only: true` option
+  * Moved `ElasticsearchEx.Api.Document.source_exists?/1` by `ElasticsearchEx.Api.Source.exists?/1`
 
 ## v0.6.2 (2024-04-01)